+++ NEWS.md
@@ -1,4 +1,4 @@
-# ambient (development version)
+# ambient 1.0.1
 
 * Added `noise_blue()` for generating blue noise texture based on 
   Void-and-cluster algorithm.